--- packages/excalidraw/bug-issue-template.js
@@ -1,11 +1,0 @@
-export default (sentryErrorId) => `
-### Scene content
-
-\`\`\`
-Paste scene content here
-\`\`\`
-
-### Sentry Error ID
-
-${sentryErrorId}
-`;
