--- packages/excalidraw/collab/index.tsx
@@ -1,5 +1,0 @@
-export { default as Collab } from "./Collab";
-
-export { Excalidraw } from "./Excalidraw";
-
-export { ExcalidrawApp } from "./ExcalidrawApp"
