+++ packages/excalidraw/package.json
@@ -1,47 +1,32 @@
 {
-  "name": "@jitsi/excalidraw",
-  "version": "0.20.0",
+  "name": "@excalidraw/excalidraw",
+  "version": "0.18.0",
   "type": "module",
-  "types": "./dist/types/packages/excalidraw/*.d.ts",
+  "types": "./dist/types/excalidraw/index.d.ts",
   "main": "./dist/prod/index.js",
   "module": "./dist/prod/index.js",
   "exports": {
     "./common/*": {
-      "types": "./dist/types/packages/common/src/*.d.ts",
-      "development": "./dist/dev/index.js",
-      "production": "./dist/prod/index.js",
-      "default": "./dist/prod/index.js"
+      "types": "./dist/types/common/src/*.d.ts"
     },
     "./element/*": {
-      "types": "./dist/types/packages/element/src/*.d.ts",
-      "development": "./dist/dev/index.js",
-      "production": "./dist/prod/index.js",
-      "default": "./dist/prod/index.js"
+      "types": "./dist/types/element/src/*.d.ts"
     },
     "./math/*": {
-      "types": "./dist/types/packages/math/src/*.d.ts",
-      "development": "./dist/dev/index.js",
-      "production": "./dist/prod/index.js",
-      "default": "./dist/prod/index.js"
+      "types": "./dist/types/math/src/*.d.ts"
     },
     "./utils/*": {
-      "types": "./dist/types/packages/utils/src/*.d.ts",
-      "development": "./dist/dev/index.js",
-      "production": "./dist/prod/index.js",
-      "default": "./dist/prod/index.js"
+      "types": "./dist/types/utils/src/*.d.ts"
     },
     "./*": {
-      "types": "./dist/types/packages/excalidraw/*.d.ts",
-      "development": "./dist/dev/index.js",
-      "production": "./dist/prod/index.js",
-      "default": "./dist/prod/index.js"
+      "types": "./dist/types/excalidraw/*.d.ts"
     },
     "./index.css": {
       "development": "./dist/dev/index.css",
       "production": "./dist/prod/index.css"
     },
     ".": {
-      "types": "./dist/types/packages/excalidraw/index.d.ts",
+      "types": "./dist/types/excalidraw/index.d.ts",
       "development": "./dist/dev/index.js",
       "production": "./dist/prod/index.js",
       "default": "./dist/prod/index.js"
@@ -60,8 +45,7 @@
     "excalidraw-embed",
     "react",
     "npm",
-    "npm jitsi",
-    "jitsi"
+    "npm excalidraw"
   ],
   "browserslist": {
     "production": [
@@ -87,10 +71,6 @@
     "react-dom": "^17.0.2 || ^18.2.0 || ^19.0.0"
   },
   "dependencies": {
-    "@excalidraw/common": "workspace:*",
-    "@excalidraw/element": "workspace:*", 
-    "@excalidraw/math": "workspace:*",
-    "@excalidraw/utils": "workspace:*",
     "@braintree/sanitize-url": "6.0.2",
     "@excalidraw/laser-pointer": "1.3.1",
     "@excalidraw/mermaid-to-excalidraw": "1.1.2",
@@ -121,12 +101,7 @@
     "pwacompat": "2.0.17",
     "roughjs": "4.6.4",
     "sass": "1.51.0",
-    "tunnel-rat": "0.1.2",
-    "socket.io-client": "4.7.2",
-    "firebase": "11.3.1",
-    "idb-keyval": "6.0.3",
-    "@sentry/browser": "9.0.1",
-    "i18next-browser-languagedetector": "6.1.4"
+    "tunnel-rat": "0.1.2"
   },
   "devDependencies": {
     "@size-limit/preset-big-lib": "9.0.0",
@@ -150,9 +125,9 @@
     "jest-diff": "29.7.0",
     "typescript": "4.9.4"
   },
-  "repository": "https://github.com/yashop7/Jitsi-excalidraw",
-  "bugs": "https://github.com/yashop7/Jitsi-excalidraw/issues",
-  "homepage": "https://github.com/yashop7/Jitsi-excalidraw/tree/master/packages/excalidraw",
+  "repository": "https://github.com/excalidraw/excalidraw",
+  "bugs": "https://github.com/excalidraw/excalidraw/issues",
+  "homepage": "https://github.com/excalidraw/excalidraw/tree/master/packages/excalidraw",
   "scripts": {
     "gen:types": "rimraf types && tsc",
     "build:esm": "rimraf dist && node ../../scripts/buildPackage.js && yarn gen:types"
