--- packages/package.json
@@ -1,11 +1,0 @@
-{
-  "workspaces": [
-    "packages/*"
-  ],
-  "dependencies": {
-    "@excalidraw/common": "workspace:*",
-    "@excalidraw/element": "workspace:*", 
-    "@excalidraw/math": "workspace:*",
-    "@excalidraw/utils": "workspace:*"
-  }
-}
